Move common functions back to async_.py
