+++ youtube_audio_matcher/download/__init__.py
@@ -1,19 +1,10 @@
-from .common import (
-    download_video_mp3, get_videos_page_url, video_metadata_from_source,
-)
-
 from .async_ import (
     async_download_channels, async_download_video_mp3s, async_get_source,
     async_video_metadata_from_urls, 
 )
 
-from .sync import (
-    download_channel, download_channels, download_video_mp3s, get_source,
-)
-
 __all__ = [
     "async_download_channels", "async_download_video_mp3s", "async_get_source",
-    "async_video_metadata_from_urls", "download_channel", "download_channels",
-    "download_video_mp3s", "get_source", "download_video_mp3",
+    "async_video_metadata_from_urls", "download_video_mp3",
     "get_videos_page_url", "video_metadata_from_source",
 ]
